requirements.txt
numpy
pandas
matplotlib
seaborn
yfinance
scikit-learn
tensorflow
keras
python-dotenv


